requirements.txt
fastapi
websocket-client
